+++ CHEWBBACA/AlleleCallEvaluator/evaluate_calls.py
@@ -28,10 +28,10 @@ try:
 		fasta_operations as fao,
 		iterables_manipulation as im,
 		multiprocessing_operations as mo,
-		distance_matrix as dm,
 		mafft_wrapper as mw,
 		fasttree_wrapper as fw)
 	from ExtractCgMLST import determine_cgmlst
+	from ComputeDistances import compute_distances
 except ModuleNotFoundError:
 	from CHEWBBACA.utils import (
 		constants as ct,
@@ -39,10 +39,10 @@ except ModuleNotFoundError:
 		fasta_operations as fao,
 		iterables_manipulation as im,
 		multiprocessing_operations as mo,
-		distance_matrix as dm,
 		mafft_wrapper as mw,
 		fasttree_wrapper as fw)
 	from CHEWBBACA.ExtractCgMLST import determine_cgmlst
+	from CHEWBBACA.ComputeDistances import compute_distances
 
 
 def compute_sample_stats(sample_ids, total_loci, coordinates_file,
@@ -373,8 +373,7 @@ def main(input_files, schema_directory, output_directory, annotations,
 			print('done.')
 			# Compute Presence-Absence matrix
 			print('Computing Presence-Absence matrix...', end='')
-			pa_matrix, pa_outfile = determine_cgmlst.presAbs(masked_profiles,
-															 output_directory)
+			pa_matrix, pa_outfile = determine_cgmlst.compute_presence_absence(masked_profiles, output_directory)
 			print('done.')
 
 			if no_pa is False:
@@ -389,8 +388,8 @@ def main(input_files, schema_directory, output_directory, annotations,
 			if no_dm is False or no_tree is False or cg_alignment is True:
 				# Compute the cgMLST at 100%
 				print('Determining cgMLST loci...')
-				cgMLST_genes, _ = determine_cgmlst.compute_cgMLST(pa_matrix, sample_ids,
-																  1, len(sample_ids))
+				cgMLST_genes, _, _, _ = determine_cgmlst.compute_cgMLST(pa_matrix, sample_ids,
+																  1, len(sample_ids), False)
 				cgMLST_genes = cgMLST_genes.tolist()
 				print('\n', f'cgMLST is composed of {len(cgMLST_genes)} loci.')
 				if len(cgMLST_genes) > 0:
@@ -402,10 +401,10 @@ def main(input_files, schema_directory, output_directory, annotations,
 				# Compute distance matrix
 				# Based on cgMLST profiles
 				if len(cgMLST_genes) > 0:
-					dm_file = dm.main(cgMLST_matrix_outfile, output_directory,
-									  cpu_cores, True, True)
+					distances_dir = fo.join_paths(output_directory, ['distances'])
+					dm_file = compute_distances.main(cgMLST_matrix_outfile, distances_dir, 'hamming', 'symmetric', True, False, cpu_cores)
 					# Import distance matrix
-					distance_m = pd.read_csv(dm_file[0], header=0, index_col=0,
+					distance_m = pd.read_csv(dm_file, header=0, index_col=0,
 											 sep='\t', low_memory=False)
 					# Convert Index values and column names to str to avoid issues with IDs that can be interpreted as int
 					distance_m.index = distance_m.index.astype('string')