+++ CHEWBBACA/docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2025, UMMI'
 author = 'Rafael Mamede'
 
 # The full version, including alpha/beta/rc tags
-release = '3.8.0'
+release = '3.9.0'
 
 # -- General configuration ---------------------------------------------------